backend/requirements.txt
fastapi
uvicorn
sqlmodel
google-generativeai
pypdf
beautifulsoup4
requests
python-multipart
pytest
httpx
pytest-asyncio

